code converter for multipart
