+++ parser/parser.go
@@ -89,6 +89,7 @@ func (p *Parser) Parse(text string) (*gabs.Container, error) {
 	p.cache = make(map[string][]string)
 	p.LineNum = 0
 	res, _ := p.Pm.Start()
+	// TODO : Getting Null value from Start() function for multistage files
 	_, err := p.assertEnd()
 	if err != nil {
 		return nil, err
